+++ playgrounds/svelte/vite.config.js
@@ -4,7 +4,7 @@ import { defineConfig } from 'vite'
 
 export default defineConfig({
   plugins: [
-    laravel.default({
+    laravel({
       input: ['resources/css/app.css', 'resources/js/app.js'],
       ssr: 'resources/js/ssr.js',
       refresh: true,